old_local_client/requirements.txt
flake8
google-cloud-translate
click
pyinstaller
PySocks
requests
openpyxl
six
Mako
flask
pycryptodome
python-dotenv
